requirements.txt
fastapi
uvicorn
sqlalchemy
mido
dotenv
python-multipart
python-jose[cryptography]
jinja2